Add duty cycle tracking to replayer/gapfiller
