+++ artio-core/src/main/java/uk/co/real_logic/artio/engine/logger/GapFiller.java
@@ -16,6 +16,7 @@
 package uk.co.real_logic.artio.engine.logger;
 
 import io.aeron.Subscription;
+import io.aeron.driver.DutyCycleTracker;
 import io.aeron.logbuffer.BufferClaim;
 import io.aeron.logbuffer.Header;
 import org.agrona.DirectBuffer;
@@ -63,9 +64,11 @@ public class GapFiller extends AbstractReplayer
         final SenderSequenceNumbers senderSequenceNumbers,
         final ReplayerCommandQueue replayerCommandQueue,
         final FixSessionCodecsFactory fixSessionCodecsFactory,
-        final EpochNanoClock clock)
+        final EpochNanoClock clock,
+        final DutyCycleTracker dutyCycleTracker)
     {
-        super(publication.dataPublication(), fixSessionCodecsFactory, new BufferClaim(), senderSequenceNumbers);
+        super(publication.dataPublication(), fixSessionCodecsFactory, new BufferClaim(), senderSequenceNumbers,
+            clock, dutyCycleTracker);
         this.inboundSubscription = inboundSubscription;
         this.publication = publication;
         this.agentNamePrefix = agentNamePrefix;
@@ -76,7 +79,10 @@ public class GapFiller extends AbstractReplayer
 
     public int doWork()
     {
-        timestamper.sendTimestampMessage();
+        final long timeInNs = clock.nanoTime();
+
+        trackDutyCycleTime(timeInNs);
+        timestamper.sendTimestampMessage(timeInNs);
 
         return replayerCommandQueue.poll() + inboundSubscription.controlledPoll(this, POLL_LIMIT);
     }
